backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gunicorn>=22.0.0
httpx>=0.27.0,<0.28
# pydantic 2.10+ и pydantic-core 2.34+ — колёса под Python 3.13 Windows, без Rust
pydantic>=2.10,<3
pydantic-core>=2.34
python-multipart==0.0.6
# psycopg2-binary: предсобранные колёса под Windows, без Rust
psycopg2-binary>=2.9.9
certifi==2024.7.4
PySocks>=1.7.1
PyYAML>=6.0
python-docx>=1.1.0
pypdf>=4.0.2
beautifulsoup4>=4.12.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
# Для cosine similarity в сравнении формулировок
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
langgraph>=0.2.0,<0.3
langchain-core>=0.3.0,<0.4
langchain-openai>=0.2.0,<0.4
json-repair>=0.30.0
redis>=5.0.0,<6
